BMI270 not init

+++ Core/Src/Quaternions.c
@@ -93,10 +93,10 @@ void filterUpdate(float gx, float gy, float gz, float ax, float ay, float az) {
 	}
 
 	// Integrate rate of change of quaternion to yield quaternion
-	q0 += qDot1 * (1.0f / (uint16_t) sampleFreq);
-	q1 += qDot2 * (1.0f / (uint16_t) sampleFreq);
-	q2 += qDot3 * (1.0f / (uint16_t) sampleFreq);
-	q3 += qDot4 * (1.0f / (uint16_t) sampleFreq);
+	q0 += qDot1 * (1.0f / sampleFreq);
+	q1 += qDot2 * (1.0f / sampleFreq);
+	q2 += qDot3 * (1.0f / sampleFreq);
+	q3 += qDot4 * (1.0f / sampleFreq);
 
 	// Normalise quaternion
 	recipNorm = invSqrt(q0 * q0 + q1 * q1 + q2 * q2 + q3 * q3);
